requirements.txt
flask
requests
pyyaml
kubernetes
kubeconfig
